refactor(job): job 패키지 분리
