+++ back-end/src/main/java/admin/gpuserver/dto/response/GpuServerResponse.java
@@ -2,7 +2,8 @@ package admin.gpuserver.dto.response;
 
 import admin.gpuserver.domain.GpuBoard;
 import admin.gpuserver.domain.GpuServer;
-import admin.gpuserver.domain.Job;
+import admin.job.domain.Job;
+import admin.job.dto.response.JobResponse;
 
 import java.util.List;
