+++ back-end/src/main/java/admin/job/domain/JobStatus.java
@@ -1,4 +1,4 @@
-package admin.gpuserver.domain;
+package admin.job.domain;
 
 public enum JobStatus {
     WAITING, RUNNING, COMPLETED, CANCELED
