+++ back-end/src/main/java/admin/job/dto/response/JobResponse.java
@@ -1,7 +1,7 @@
-package admin.gpuserver.dto.response;
+package admin.job.dto.response;
 
-import admin.gpuserver.domain.Job;
-import admin.gpuserver.domain.JobStatus;
+import admin.job.domain.Job;
+import admin.job.domain.JobStatus;
 
 import java.util.List;
 import java.util.stream.Collectors;
